refactor: merge PDF and Excel exports into unified ExportService
