+++ salary-service/src/main/java/com/micro/salaryservice/service/ExportService.java
@@ -2,6 +2,10 @@ package com.micro.salaryservice.service;
 
 import jakarta.servlet.http.HttpServletResponse;
 
-public interface ExcelService {
+import java.io.IOException;
+
+public interface ExportService {
     Object exportExcelFile(HttpServletResponse response);
+
+    void exportPdfFile(HttpServletResponse response) throws IOException;
 }
