--- salary-service/src/main/java/com/micro/salaryservice/service/PdfService.java
@@ -1,9 +1,0 @@
-package com.micro.salaryservice.service;
-
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.io.IOException;
-
-public interface PdfService {
-    void exportPdfFile(HttpServletResponse response) throws IOException;
-}
